# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "description": "Trả JWT; gửi ở header `Authorization: Bearer <accessToken>` cho mọi API khác. Sai tên đăng nhập hoặc mật khẩu cùng một thông báo 401. Tài khoản/người dùng bị khóa cũng trả 401.\n\n**Quyền:** Công khai, không cần token.",
        "operationId": "Auth_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu gửi lên không hợp lệ (message là mảng lỗi validate) hoặc vi phạm CHECK của CSDL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Thiếu/sai/hết hạn JWT, hoặc tài khoản/người dùng đang bị khóa hay ngừng hoạt động",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Đăng nhập",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "description": "Lấy từ JWT (không có `hoTen`); `exp` cho biết token hết hạn lúc nào.\n\n**Quyền:** mọi tài khoản đã đăng nhập.",
        "operationId": "Auth_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JwtUserDto"
                }
              }
            }
          },
          "401": {
            "description": "Thiếu/sai/hết hạn JWT, hoặc tài khoản/người dùng đang bị khóa hay ngừng hoạt động",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Thông tin tài khoản đang đăng nhập",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "QLTV Nhóm 8 - API quản lý thư viện",
    "description": "API quản lý thư viện đại học (đồ án IE103, nhóm 8). Nghiệp vụ nằm trong MySQL (procedure, trigger, view); BE chỉ gọi lại và xác thực.\n\n**Cách dùng nhanh**\n1. `POST /auth/login` với tài khoản seed (ví dụ `ad001` / `AD001@Nhom8`), lấy `accessToken`.\n2. Bấm **Authorize** ở góc phải và dán token (không cần gõ chữ `Bearer`). Token được nhớ khi tải lại trang.\n3. Mỗi endpoint ghi rõ **Quyền** được gọi. Sai vai trò: 403. Token sai, hết hạn hoặc tài khoản bị khóa: 401.\n\n**Quy ước dữ liệu**\n- Khóa chính BIGINT (`id`) là **chuỗi**, ví dụ `\"13\"`. Các mã (`maSach`, `maBanSach`, `maNguoiDung`...) là chuỗi và là thứ dùng khi gọi API nghiệp vụ.\n- Endpoint danh sách có phân trang trả `{ data, total, page, limit }` (`page` từ 1, `limit` tối đa 100).\n- Cột ngày kiểu DATE trả dạng ISO `YYYY-MM-DDT00:00:00.000Z`.\n- Tiền (VND) từ báo cáo và `/me` là number; `soTien`/`giaBia` đọc từ bảng là chuỗi thập phân (Prisma Decimal).\n- Báo cáo (`/bao-cao/*`), `GET /sach` và `/me/*` giữ tên cột snake_case của view; các endpoint còn lại dùng camelCase.\n- Lỗi luôn có dạng `{ statusCode, message }`. 422 là lỗi nghiệp vụ do CSDL báo, `message` là thông báo cho người dùng (tiếng Việt không dấu).",
    "version": "2.2",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "tenDangNhap": {
            "type": "string",
            "maxLength": 80,
            "example": "ad001"
          },
          "matKhau": {
            "type": "string",
            "maxLength": 72,
            "example": "AD001@Nhom8"
          }
        },
        "required": [
          "tenDangNhap",
          "matKhau"
        ]
      },
      "VaiTroTaiKhoan": {
        "type": "string",
        "enum": [
          "ADMIN",
          "THU_THU",
          "BAN_DOC"
        ]
      },
      "LoginUserDto": {
        "type": "object",
        "properties": {
          "nguoiDungId": {
            "type": "string",
            "example": "13",
            "description": "nguoi_dung.id (BIGINT dạng chuỗi)"
          },
          "maNguoiDung": {
            "type": "string",
            "example": "AD001",
            "description": "Mã người dùng (dùng làm tham số maNguoiDung ở các API khác)"
          },
          "tenDangNhap": {
            "type": "string",
            "example": "ad001"
          },
          "vaiTro": {
            "nullable": false,
            "allOf": [
              {
                "$ref": "#/components/schemas/VaiTroTaiKhoan"
              }
            ]
          },
          "hoTen": {
            "type": "string",
            "example": "Quản trị hệ thống"
          }
        },
        "required": [
          "nguoiDungId",
          "maNguoiDung",
          "tenDangNhap",
          "vaiTro",
          "hoTen"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT: gửi ở header `Authorization: Bearer <accessToken>`"
          },
          "user": {
            "$ref": "#/components/schemas/LoginUserDto"
          }
        },
        "required": [
          "accessToken",
          "user"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 422
          },
          "error": {
            "type": "string",
            "example": "Unprocessable Entity",
            "description": "Tên mã HTTP (chỉ có ở lỗi do bộ lọc CSDL và một số lỗi của Nest)"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Nguoi dung dang no phat",
            "description": "Chuỗi thông báo; lỗi validate dữ liệu gửi lên (400) là mảng các thông báo"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
      "JwtUserDto": {
        "type": "object",
        "properties": {
          "nguoiDungId": {
            "type": "string",
            "example": "13",
            "description": "nguoi_dung.id (BIGINT dạng chuỗi)"
          },
          "maNguoiDung": {
            "type": "string",
            "example": "AD001",
            "description": "Mã người dùng (dùng làm tham số maNguoiDung ở các API khác)"
          },
          "tenDangNhap": {
            "type": "string",
            "example": "ad001"
          },
          "vaiTro": {
            "nullable": false,
            "allOf": [
              {
                "$ref": "#/components/schemas/VaiTroTaiKhoan"
              }
            ]
          },
          "iat": {
            "type": "integer",
            "example": 1791360000,
            "description": "Thời điểm cấp token (giây Unix)"
          },
          "exp": {
            "type": "integer",
            "example": 1791388800,
            "description": "Thời điểm token hết hạn (giây Unix); sau đó mọi API trả 401"
          }
        },
        "required": [
          "nguoiDungId",
          "maNguoiDung",
          "tenDangNhap",
          "vaiTro",
          "iat",
          "exp"
        ]
      },
...
    }
  }
}